+++ pkg/db/implicit_repr.go
@@ -1,11 +1,8 @@
 package db
 
 import (
-	"database/sql"
-	"encoding/hex"
 	"fmt"
 	"math/big"
-	"math/rand"
 	"strings"
 )
 
@@ -42,168 +39,9 @@ func (n *node) FullID() [33]byte {
 	return id
 }
 
-// DeletemeCreateNodes2 where cound is the number of leaves
-func DeletemeCreateNodes2(db DB, count int) error {
-	var err error
-	c := db.(*mysqlDB)
-
-	root := &node{
-		id:    big.NewInt(0),
-		depth: 0,
-	}
-	uniqueLeaves := make(map[[32]byte]struct{})
-	for i := 0; i < count; i++ {
-		var idhash [32]byte
-		if _, err = rand.Read(idhash[:]); err != nil {
-			return err
-		}
-		if _, ok := uniqueLeaves[idhash]; ok {
-			panic("duplicate random ID")
-		}
-		uniqueLeaves[idhash] = struct{}{}
-		updateStructureRaw(root, idhash)
-	}
-	dups := findDuplicates(root) // deleteme
-	if len(dups) > 0 {
-		fmt.Printf("%d duplicates found\n", len(dups))
-		for id, d := range dups {
-			fmt.Printf("ID: [%s] %2d nodes\n", hex.EncodeToString(id[:]), len(d))
-			for i, c := range d {
-				fmt.Printf("\t[%2d] depth %d\n\n", i, c.depth)
-				tempId := c.FullID()
-				fmt.Printf("\thex: %s\n\tbits: %s\n",
-					hex.EncodeToString(tempId[1:]), bitString(c.id))
-				fmt.Println(pathToString(pathFromNode(c)))
-			}
-		}
-		panic("duplicates")
-	}
-	if err = insertIntoDB(c, root); err != nil {
-		return err
-	}
-	return nil
-}
-
 func updateStructureRaw(root *node, leafHash [32]byte) {
 	id := big.NewInt(0).SetBytes(leafHash[:])
 	updateStructure(root, id)
-}
-
-func insertIntoDB(c *mysqlDB, root *node) error {
-	// var err error
-	// _, err = c.db.Exec("LOCK TABLES nodes WRITE;")
-	// if err != nil {
-	// 	return err
-	// }
-	// _, err = c.db.Exec("SET autocommit=0") // XXX(juagargi) this seems to cause some trouble by omitting some records
-	// if err != nil {
-	// 	return err
-	// }
-	// _, err = c.db.Exec("ALTER TABLE nodes DROP INDEX idhash; ;")
-	// if err != nil {
-	// 	return err
-	// }
-
-	N := 1000
-
-	repeatedStmt := "INSERT INTO nodes (idhash,parentnode,leftnode,rightnode,value) VALUES " + repeatStmt(N, 5)
-	stmt, err := c.db.Prepare(repeatedStmt)
-	if err != nil {
-		panic(err)
-	}
-	pending := make([]*node, 0) // XXX(juagargi) !!!! this queue is not memory efficient
-	// the root node will not be inserted into the table (performance reasons), and thus, its
-	// two children cannot point to it. They will point to null.
-	if root.left != nil {
-		root.left.parent = nil
-		pending = append(pending, root.left)
-	}
-	if root.right != nil {
-		root.right.parent = nil
-		pending = append(pending, root.right)
-	}
-	total := 0
-	nodes := make([]*node, 0, N)
-	for i := 0; len(pending) > 0; i++ {
-		total++
-		if total%10000 == 0 {
-			fmt.Printf("inserted %7d\n", total)
-		}
-		current := pending[0]
-		pending = pending[1:]
-		if current.left != nil {
-			pending = append(pending, current.left)
-		}
-		if current.right != nil {
-			pending = append(pending, current.right)
-		}
-
-		if len(nodes) == N {
-			// send to DB
-			if err := nodeBatchToDB(c, stmt, nodes); err != nil {
-				return err
-			}
-			// clear
-			nodes = make([]*node, 0, N)
-			i = -1
-		}
-		nodes = append(nodes, current)
-	}
-	// remaining less than N-1 ones:
-	if len(nodes) > 0 {
-		repeatedStmt = "INSERT INTO nodes (idhash,parentnode,leftnode,rightnode,value) VALUES " +
-			repeatStmt(len(nodes), 5)
-		stmt, err = c.db.Prepare(repeatedStmt)
-		if err != nil {
-			panic(err)
-		}
-		if err := nodeBatchToDB(c, stmt, nodes); err != nil {
-			return err
-		}
-	}
-
-	// _, err = c.db.Exec("COMMIT")
-	// if err != nil {
-	// 	return err
-	// }
-	// _, err = c.db.Exec("ALTER TABLE nodes ADD UNIQUE INDEX idhash (idhash ASC) VISIBLE; ;")
-	// if err != nil {
-	// 	return err
-	// }
-	// _, err = c.db.Exec("UNLOCK TABLES")
-	// if err != nil {
-	// 	return err
-	// }
-	return nil
-}
-
-// nodeBatchToDB creates the data structure to be pushed to the DB, from a slice of nodes.
-// It expects the root node not to be present in the collection.
-func nodeBatchToDB(c *mysqlDB, stmt *sql.Stmt, nodes []*node) error {
-	toSlice := func(id [33]byte) []byte {
-		return id[:]
-	}
-	cryptoMaterialMock, _ := hex.DecodeString("deadbeef")
-	data := make([]interface{}, 5*len(nodes))
-	for i := 0; i < len(nodes); i++ {
-		n := nodes[i]
-		data[5*i] = toSlice(n.FullID())
-		// fmt.Printf("%4d ID: %s\n", i, hex.EncodeToString(toSlice(n.FullID())))
-		if n.parent != nil {
-			data[5*i+1] = toSlice(n.parent.FullID())
-		}
-		if n.left != nil {
-			data[5*i+2] = toSlice(n.left.FullID())
-		}
-		if n.right != nil {
-			data[5*i+3] = toSlice(n.right.FullID())
-		}
-		data[5*i+4] = cryptoMaterialMock
-	}
-	if _, err := stmt.Exec(data...); err != nil {
-		return fmt.Errorf("executing prep. statement: %w", err)
-	}
-	return nil
 }
 
 func updateStructure(root *node, leafId *big.Int) {
